% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{Data}
\alias{Data}
\title{Data Class}
\description{
Data description goes here.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connectionDetails}}{(\verb{[DatabaseConnector][DatabaseConnector::connectionDetails]})\cr
Connection details.}

\item{\code{cdmDatabaseSchema}}{(\code{character(1)})\cr
CDM database schema name.}

\item{\code{cohortDatabaseSchema}}{(\code{character(1)})\cr
Database schema with write access to create scratch tables.}

\item{\code{cohortTable}}{(\code{character(1)})\cr
Table name where the target cohorts will be generated.}

\item{\code{outcomeDatabaseSchema}}{(\code{character(1)})\cr
Outcome database schema with write access to create scratch tables.}

\item{\code{outcomeTable}}{(\code{character(1)})\cr
Table name where the outcome cohorts will be generated.}

\item{\code{oracleTempSchema}}{(\code{character(1)})\cr
Oracle temp schema.}

\item{\code{sampleSize}}{(\code{numeric(1)})\cr
Sample size for the cohort query.}

\item{\code{cdmVersion}}{(\code{numeric(1)})\cr
CDM version number of the database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Data$new()}}
\item \href{#method-createCohorts}{\code{Data$createCohorts()}}
\item \href{#method-clone}{\code{Data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-createCohorts}{}}}
\subsection{Method \code{createCohorts()}}{
Creates the cohorts in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$createCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema,
  outputFolder
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connectionDetails}}{(\code{list(1)})\cr
An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{\code{cdmDatabaseSchema}}{(\code{character(1)})\cr
Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{\code{cohortDatabaseSchema}}{(\code{character(1)})\cr
Schema name where intermediate data can be stored. You will need to have
write privileges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{\code{cohortTable}}{(\code{character(1)})\cr
The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{\code{oracleTempSchema}}{(\code{character(1)})\cr
Should be used in Oracle to specify a schema where the user has write
privileges for storing temporary tables.}

\item{\code{outputFolder}}{(\code{character(1)})\cr
Name of local folder to place results; make sure to use forward slashes (/).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
