% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Model Class}
\description{
Model class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{(\code{character(1)})\cr
Unique name of the model.}

\item{\code{cohortId}}{(\code{numeric(1)})\cr
Target cohort identifier.}

\item{\code{outcomeId}}{(\code{numeric(N)})\cr
List of outcome cohort identifiers.}

\item{\code{attr_predictionType}}{(\code{character(1)})\cr
Prediction type.}

\item{\code{attr_type}}{(\code{character(1)})\cr
Model type.}

\item{\code{model}}{(\verb{[Parameter]})\cr
Single set of model parameters}

\item{\code{populationSettings}}{(\code{TBD})\cr
Population settings.}

\item{\code{covariateSettings}}{(\code{TBD})\cr
Covariate settings.}

\item{\code{covariateMap}}{(\code{TBD})\cr
Covariate map.}

\item{\code{varImp}}{(\code{TBD})\cr
Variable importance.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Model$new()}}
\item \href{#method-predict}{\code{Model$predict()}}
\item \href{#method-loadModelFromJson}{\code{Model$loadModelFromJson()}}
\item \href{#method-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(
  cohortId = NULL,
  outcomeId = NULL,
  coefficients = NULL,
  finalMapping = "function(x){return(x)}",
  offset = 0,
  baselineHazard = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cohortId}}{(\code{numeric(1)})\cr
Target cohort identifier.}

\item{\code{outcomeId}}{(\code{numeric(N)})\cr
Vector of outcome cohort identifiers.}

\item{\code{coefficients}}{(\code{data.frame(1)})\cr
Model coefficients.}

\item{\code{finalMapping}}{(\code{character(1)})\cr
Mapping function of the form 'function(x){return(x)}'.}

\item{\code{offset}}{(\code{numeric(1)})\cr
Offset.}

\item{\code{baselineHazard}}{(\code{numeric(1)})\cr
Baseline hazard.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
predict
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$predict()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadModelFromJson"></a>}}
\if{latex}{\out{\hypertarget{method-loadModelFromJson}{}}}
\subsection{Method \code{loadModelFromJson()}}{
loadModelFromJson
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$loadModelFromJson()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
