% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter.R
\name{Parameter}
\alias{Parameter}
\title{Model Class}
\description{
Model description goes here.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{coefficients}}{(\code{data.frame(1)})\cr
Model coefficients.}

\item{\code{finalMapping}}{(\code{character(1)})\cr
Mapping function of the form 'function(x){return(x)}'.}

\item{\code{offset}}{(\code{numeric(1)})\cr
Offset.}

\item{\code{baselineHazard}}{(\code{numeric(1)})\cr
Baseline hazard.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Parameter$new()}}
\item \href{#method-setCoefficients}{\code{Parameter$setCoefficients()}}
\item \href{#method-getParameters}{\code{Parameter$getParameters()}}
\item \href{#method-loadParameterFromJson}{\code{Parameter$loadParameterFromJson()}}
\item \href{#method-clone}{\code{Parameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$new(
  coefficients,
  finalMapping = "function(x){return(x)}",
  offset = numeric(),
  baselineHazard = numeric()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{coefficients}}{(\code{data.frame(1)})\cr
Model coefficients.}

\item{\code{finalMapping}}{(\code{character(1)})\cr
Mapping function of the form 'function(x){return(x)}'.}

\item{\code{offset}}{(\code{numeric(1)})\cr
Offset.}

\item{\code{baselineHazard}}{(\code{numeric(1)})\cr
Baseline hazard.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setCoefficients"></a>}}
\if{latex}{\out{\hypertarget{method-setCoefficients}{}}}
\subsection{Method \code{setCoefficients()}}{
Sets the coefficients.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$setCoefficients(covariateSettings)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariateSettings}}{(\code{list(1)} or \code{data.frame(1)}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getParameters"></a>}}
\if{latex}{\out{\hypertarget{method-getParameters}{}}}
\subsection{Method \code{getParameters()}}{
getParameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$getParameters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadParameterFromJson"></a>}}
\if{latex}{\out{\hypertarget{method-loadParameterFromJson}{}}}
\subsection{Method \code{loadParameterFromJson()}}{
Load the parameters from a JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$loadParameterFromJson(path = NULL, modelName = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(\code{character(1)}).}

\item{\code{modelName}}{(\code{character(1)})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Parameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
