% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study.R
\name{Study}
\alias{Study}
\title{Study Class}
\description{
Model description goes here.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{packageName}}{(\code{character(1)})\cr
Package name.}

\item{\code{packageDescription}}{(\code{character(1)})\cr
Package description.}

\item{\code{createdBy}}{(\code{character(1)})\cr
Created by names.}

\item{\code{organizationName}}{(\code{character(1)})\cr
Organization name.}

\item{\code{analysisSettings}}{(\code{data.frame(1)})\cr
Analysis settings.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Study$new()}}
\item \href{#method-loadStudyFromJson}{\code{Study$loadStudyFromJson()}}
\item \href{#method-clone}{\code{Study$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$new(studyPath = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studyPath}}{(\code{character(1)})\cr
Path to a prediction study in the JSON format.}

\item{\code{studyPath}}{(\code{character(1)})\cr
Path to a prediction study in the JSON format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadStudyFromJson"></a>}}
\if{latex}{\out{\hypertarget{method-loadStudyFromJson}{}}}
\subsection{Method \code{loadStudyFromJson()}}{
Load study from JSON file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$loadStudyFromJson(studyPath = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{studyPath}}{(\code{character(1)})\cr
Path to a prediction study in the JSON format.}

\item{\code{studyPath}}{(\code{character(1)})\cr
Path to a prediction study in the JSON format.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Study$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
